Extract tokens
